Conflicts:
	Source/Core/BareMetal/L0Types/Architecture/x86_cl/CompilerTypes.h
	Source/Core/BareMetal/L0Types/LinkedListHolder.cpp
	Source/Core/BareMetal/L1Portability/Architecture/x86_cl/HighResolutionTimerA.h
	Source/Core/BareMetal/L1Portability/Architecture/x86_gcc/HighResolutionTimerA.h
	Source/Core/BareMetal/L1Portability/Environment/Generic/MemoryOperationsHelper_CLIB_Generic.cpp
	Source/Core/BareMetal/L1Portability/Environment/Linux/BasicConsole.cpp
	Source/Core/BareMetal/L1Portability/Environment/Linux/HighResolutionTimer.h
	Source/Core/BareMetal/L1Portability/Environment/Linux/HighResolutionTimerCalibrator.cpp
	Source/Core/BareMetal/L1Portability/Environment/Linux/HighResolutionTimerCalibrator.h
	Source/Core/BareMetal/L1Portability/Environment/Linux/Sleep.cpp
	Source/Core/BareMetal/L1Portability/Environment/Linux/lint-linux.h
	Source/Core/BareMetal/L1Portability/Environment/Windows/HighResolutionTimerCalibrator.cpp
	Source/Core/BareMetal/L1Portability/Environment/Windows/HighResolutionTimerCalibrator.h
	Source/Core/BareMetal/L1Portability/ErrorInformation.h
	Source/Core/BareMetal/L1Portability/FastPollingMutexSem.cpp
	Source/Core/BareMetal/L1Portability/FastPollingMutexSem.h
	Source/Core/BareMetal/L1Portability/HighResolutionTimer.h
	Source/Core/BareMetal/L1Portability/Sleep.h
	Source/Core/BareMetal/L1Portability/StaticList.h
	Source/Core/BareMetal/L1Portability/TimeoutType.h
	Source/Core/BareMetal/L2Objects/Object.cpp
	Source/Core/BareMetal/L2Objects/Object.h
	Source/Core/BareMetal/L2Objects/ObjectRegistryDatabase.cpp
	Source/Core/BareMetal/L2Objects/ObjectRegistryDatabase.h
	Source/Core/BareMetal/L2Objects/Reference.cpp
	Source/Core/BareMetal/L2Objects/Reference.h
	Source/Core/BareMetal/L2Objects/ReferenceContainer.cpp
	Source/Core/BareMetal/L2Objects/ReferenceContainer.h
	Source/Core/BareMetal/L2Objects/ReferenceT.h
	Source/Core/BareMetal/L2Objects/StructuredDataI.h

+++ Source/Core/FileSystem/L1Portability/Environment/Linux/BasicFile.cpp
@@ -320,7 +320,7 @@ bool BasicFile::Read(char8* const output,
                      uint32 & size) {
     bool retVal = true;
     if (CanRead()) {
-        size = static_cast<uint32>(read(properties.identifier, output, static_cast<uint64>(size)));
+        size = static_cast<uint32>(read(properties.identifier, output, static_cast<size_t>(size)));
         retVal = (size != 0xFFFFFFFFU);
         if (retVal == false) {
             REPORT_ERROR(ErrorManagement::FatalError, "BasicFile::Read().File cannot be read");
@@ -378,7 +378,7 @@ bool BasicFile::Write(const char8 * const input,
                       uint32 & size) {
     bool retVal = true;
     if (CanWrite()) {
-        size = static_cast<uint32>(write(properties.identifier, input, static_cast<uint64>(size)));
+        size = static_cast<uint32>(write(properties.identifier, input, static_cast<size_t>(size)));
         if (size == 0xFFFFFFFFU) {
             REPORT_ERROR(ErrorManagement::FatalError, "BasicFile::Write(). File cannot be written");
             retVal = false;
@@ -525,7 +525,7 @@ bool BasicFile::SetSize(const uint64 size) {
         if (size >= static_cast<uint64>(MAX_INT64)) {
             REPORT_ERROR(ErrorManagement::FatalError, "BasicFile::SetSize(). The size is too large");
         }
-        ret = ftruncate(properties.identifier, static_cast<int64>(size));
+        ret = ftruncate(properties.identifier, static_cast<off_t>(size));
         if (ret == -1) {
             REPORT_ERROR(ErrorManagement::FatalError, "BasicFile::SetSize(). The size cannot be set");
             retVal = false;
